refactor: LottoTickets 로또 티켓 생성 및 판별 메서드 수정

+++ src/main/java/lotto/domain/LottoResult.java
@@ -5,9 +5,9 @@ import java.util.Map;
 
 public class LottoResult {
 
-    private final Map<Rank, Integer> ranks;
+    private final Map<Rank, Long> ranks;
 
-    public LottoResult(Map<Rank, Integer> ranks) {
+    public LottoResult(Map<Rank, Long> ranks) {
         this.ranks = ranks;
     }
 
@@ -21,7 +21,7 @@ public class LottoResult {
         return Math.floor((double) totalPrizeMoney / price * 100) / 100.0;
     }
 
-    public Map<Rank, Integer> getRanks() {
+    public Map<Rank, Long> getRanks() {
         return Collections.unmodifiableMap(ranks);
     }
 }
